backend/requirements.txt
flask
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
redis
flask_cors
pinecone-client
marshmallow
langchain
langchain-community
openai
tiktoken
langchain-openai
langchain-pinecone
gunicorn
pytest